+++ src/app/shared/model/counter.ts
@@ -13,8 +13,7 @@ export class Collaborator {
     public userId: string,
     public canRead: boolean = false,
     public canWrite: boolean = false,
-    public canShare: boolean = false,
-    public canDelete: boolean = false
+    public canShare: boolean = false
   ) {}
 }
 
@@ -29,7 +28,7 @@ export class Count {
 
 export class Counter {
   constructor(
-    public id: string,
+    public counterId: string,
     public name: string,
     public totalCount: number,
     public counts: Array<Count>,